refactor: merge client dir
