+++ src/client/mod.rs
@@ -1,1 +1,2 @@
 pub mod discord;
+pub mod openai;
