+++ src/client/openai.rs
